Reorganize matching logic for incoming and existing records
