+++ seed/views/properties.py
@@ -16,7 +16,7 @@ from rest_framework.decorators import detail_route, list_route
 from rest_framework.renderers import JSONRenderer
 from rest_framework.viewsets import GenericViewSet
 
-from seed.data_importer.match import merge_in_cycle_matches
+from seed.utils.match import match_merge_in_cycle
 from seed.data_importer.views import ImportFileViewSet
 from seed.decorators import ajax_request_class
 from seed.filtersets import PropertyViewFilterSet, PropertyStateFilterSet
@@ -465,7 +465,7 @@ class PropertyViewSet(GenericViewSet, ProfileIdMixin):
 
         merged_state = merge_properties(state_ids, organization_id, 'Manual Match')
 
-        count, view_id = merge_in_cycle_matches(merged_state.propertyview_set.first().id, 'PropertyState')
+        count, view_id = match_merge_in_cycle(merged_state.propertyview_set.first().id, 'PropertyState')
 
         result = {
             'status': 'success'
@@ -980,7 +980,7 @@ class PropertyViewSet(GenericViewSet, ProfileIdMixin):
                         # save the property view so that the datetime gets updated on the property.
                         property_view.save()
 
-                        count, view_id = merge_in_cycle_matches(property_view.id, 'PropertyState')
+                        count, view_id = match_merge_in_cycle(property_view.id, 'PropertyState')
 
                         if view_id is not None:
                             result.update({
@@ -1021,7 +1021,7 @@ class PropertyViewSet(GenericViewSet, ProfileIdMixin):
                         # save the property view so that the datetime gets updated on the property.
                         property_view.save()
 
-                        count, view_id = merge_in_cycle_matches(property_view.id, 'PropertyState')
+                        count, view_id = match_merge_in_cycle(property_view.id, 'PropertyState')
 
                         if view_id is not None:
                             result.update({
