+++ seed/views/taxlots.py
@@ -16,7 +16,7 @@ from rest_framework.decorators import detail_route, list_route
 from rest_framework.renderers import JSONRenderer
 from rest_framework.viewsets import GenericViewSet
 
-from seed.data_importer.match import merge_in_cycle_matches
+from seed.utils.match import match_merge_in_cycle
 from seed.data_importer.views import ImportFileViewSet
 from seed.decorators import ajax_request_class
 from seed.lib.superperms.orgs.decorators import has_perm_class
@@ -283,7 +283,7 @@ class TaxLotViewSet(GenericViewSet, ProfileIdMixin):
 
         merged_state = merge_taxlots(state_ids, organization_id, 'Manual Match')
 
-        count, view_id = merge_in_cycle_matches(merged_state.taxlotview_set.first().id, 'TaxLotState')
+        count, view_id = match_merge_in_cycle(merged_state.taxlotview_set.first().id, 'TaxLotState')
 
         result = {
             'status': 'success'
@@ -740,7 +740,7 @@ class TaxLotViewSet(GenericViewSet, ProfileIdMixin):
                         # save the property view so that the datetime gets updated on the property.
                         taxlot_view.save()
 
-                        count, view_id = merge_in_cycle_matches(taxlot_view.id, 'TaxLotState')
+                        count, view_id = match_merge_in_cycle(taxlot_view.id, 'TaxLotState')
 
                         if view_id is not None:
                             result.update({
@@ -779,7 +779,7 @@ class TaxLotViewSet(GenericViewSet, ProfileIdMixin):
                         # save the property view so that the datetime gets updated on the property.
                         taxlot_view.save()
 
-                        count, view_id = merge_in_cycle_matches(taxlot_view.id, 'TaxLotState')
+                        count, view_id = match_merge_in_cycle(taxlot_view.id, 'TaxLotState')
 
                         if view_id is not None:
                             result.update({
